Accounted for KSLabel repeat
Sometimes there are multiple Kilosort files with the column "KSLabel". I think this is before the user has manually assigned groups, and they are identical. In this case, one needs to be named 'KSLabel' in the extracted DataFrame, rather than naming them 'KSLabel_x' and _y, or the keep_good_only logic won't work.