Update isolation_distance.rst
This will need some review. without would not be used in this case but the choices are not within or outside of. I went with outside of, but if you want to be more parallel in the writing then it could be switched to spikes not within cluster `C`. 

I also changed also uses of unit to cluster to be consistent. Mathematically these are just spikes separated into clusters generated from some vector of PCs and I would say units are our neuroscientific terminology of what the cluster is (ie it is one neural unit). So we are trying to determine a good unit, but finding that it is a well isolated cluster. 

But happy to discuss if you think this is the wrong way to make this consistent.